Tournaments Refactoring
